+++ src/flexible-text/examples/module.ts
@@ -1,19 +1,21 @@
 // import { ThyTooltipModule } from './../../tooltip/tooltip.module';
 // import { ThyFlexibleTextModule } from './../flexible-text.module';
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { ThyEmptyModule, ThyTranslate, ThyFlexibleTextModule, ThyTooltipModule, ThyLabelModule } from 'ngx-tethys';
-import { TranslateModule, TranslateService } from '@ngx-translate/core';
+import { TranslateModule } from '@ngx-translate/core';
+import { ThyEmptyModule, ThyFlexibleTextModule, ThyLabelModule, ThyTooltipModule } from 'ngx-tethys';
 import { ThyFlexibleTextBasicExampleComponent } from './basic/basic.component';
-import { ThyFlexibleTextTemplateExampleComponent } from './template/template.component';
-import { ThyFlexibleTextPlacementExampleComponent } from './placement/placement.component';
 import { ThyFlexibleTextContainerExampleComponent } from './container/container.component';
+import { ThyFlexibleTextMultipleLineExampleComponent } from './multiple-line/multiple-line.component';
+import { ThyFlexibleTextPlacementExampleComponent } from './placement/placement.component';
+import { ThyFlexibleTextTemplateExampleComponent } from './template/template.component';
 const COMPONENTS = [
     ThyFlexibleTextBasicExampleComponent,
     ThyFlexibleTextTemplateExampleComponent,
     ThyFlexibleTextPlacementExampleComponent,
-    ThyFlexibleTextContainerExampleComponent
+    ThyFlexibleTextContainerExampleComponent,
+    ThyFlexibleTextMultipleLineExampleComponent
 ];
 
 @NgModule({