+++ src/input/examples/autofocus/autofocus.component.ts
@@ -1,0 +1,13 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'thy-input-autofocus-example',
+    templateUrl: './autofocus.component.html'
+})
+export class ThyInputAutofocusExampleComponent implements OnInit {
+    autofocus = false;
+
+    constructor() {}
+
+    ngOnInit() {}
+}