+++ src/input/examples/pre-suffix/pre-suffix.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'thy-input-pre-suffix-example',
+    templateUrl: './pre-suffix.component.html'
+})
+export class ThyInputPreSuffixExampleComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit() {}
+}