+++ src/table/examples/draggable/draggable.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'thy-table-draggable-example',
+    templateUrl: './draggable.component.html'
+})
+export class ThyTableDraggableExampleComponent implements OnInit {
+    data = [
+        { id: 1, name: 'Peter', age: 25, job: 'Engineer', address: 'Beijing Dong Sheng Technology' },
+        { id: 2, name: 'James', age: 26, job: 'Designer', address: 'Xian Economic Development Zone' },
+        { id: 3, name: 'Tom', age: 30, job: 'Engineer', address: 'New Industrial Park, Shushan, Hefei, Anhui' },
+        { id: 4, name: 'Elyse', age: 31, job: 'Engineer', address: 'Yichuan Ningxia' },
+        { id: 5, name: 'Jill', age: 22, job: 'DevOps', address: 'Hangzhou' }
+    ];
+
+    constructor() {}
+
+    ngOnInit(): void {}
+}