+++ CHANGELOG.md
@@ -2,6 +2,91 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="12.1.14"></a>
+## [12.1.14](https://github.com/atinc/ngx-tethys/compare/12.1.13...12.1.14) (2022-03-28)
+
+
+### Bug Fixes
+
+* **layout:** layout sidebar folder add collapse tip value #WIK-6133 ([#1551](https://github.com/atinc/ngx-tethys/issues/1551)) ([1e2fbb6](https://github.com/atinc/ngx-tethys/commit/1e2fbb6)), closes [#WIK-6133](https://github.com/atinc/ngx-tethys/issues/WIK-6133)
+
+
+
+<a name="12.1.13"></a>
+## [12.1.13](https://github.com/atinc/ngx-tethys/compare/12.1.12...12.1.13) (2022-03-28)
+
+
+### Bug Fixes
+
+* **flexible-text:** support multiline text (#INFR-2798) ([ba7e867](https://github.com/atinc/ngx-tethys/commit/ba7e867)), closes [#INFR-2798](https://github.com/atinc/ngx-tethys/issues/INFR-2798)
+
+
+### Performance Improvements
+
+* **table:** do not run change detection `pointerdown`, `touchstart` and `mousedown` events ([#1529](https://github.com/atinc/ngx-tethys/issues/1529)) ([e55a63d](https://github.com/atinc/ngx-tethys/commit/e55a63d))
+* **tree:** do not run change detection when the tree node is hovered and the drag icon visibility is changed ([ed8ea5c](https://github.com/atinc/ngx-tethys/commit/ed8ea5c))
+
+
+
+<a name="12.1.12"></a>
+## [12.1.12](https://github.com/atinc/ngx-tethys/compare/12.1.11...12.1.12) (2022-03-25)
+
+
+### Bug Fixes
+
+* **cascader:** #INFR-2781 should  scroll to active item when menu open ([#1540](https://github.com/atinc/ngx-tethys/issues/1540)) ([cbfd95f](https://github.com/atinc/ngx-tethys/commit/cbfd95f)), closes [#INFR-2781](https://github.com/atinc/ngx-tethys/issues/INFR-2781)
+* **fullscreen:** complete the `afterExited` since it has subscribers and do not run change detection on `keydown` events ([25a5ec5](https://github.com/atinc/ngx-tethys/commit/25a5ec5))
+* **layout:** adjust layout folder icon color,size and shadow #INFR-2793 ([#1543](https://github.com/atinc/ngx-tethys/issues/1543)) ([38468ca](https://github.com/atinc/ngx-tethys/commit/38468ca)), closes [#INFR-2793](https://github.com/atinc/ngx-tethys/issues/INFR-2793)
+* **resizable:** solve the style compatibility problem when dragging #INFR-2786 ([#1539](https://github.com/atinc/ngx-tethys/issues/1539)) ([e8213cc](https://github.com/atinc/ngx-tethys/commit/e8213cc)), closes [#INFR-2786](https://github.com/atinc/ngx-tethys/issues/INFR-2786)
+* **time-picker:** teardown the subscription once the time picker is destroyed ([99cfe6d](https://github.com/atinc/ngx-tethys/commit/99cfe6d))
+* **uploader:** cleanup xhr event listeners once the `upload` is unsubscribed ([590c3b9](https://github.com/atinc/ngx-tethys/commit/590c3b9))
+
+
+### Performance Improvements
+
+* **table:** do not run change detection when adding drag preview class ([a542899](https://github.com/atinc/ngx-tethys/commit/a542899))
+
+
+
+<a name="12.1.11"></a>
+## [12.1.11](https://github.com/atinc/ngx-tethys/compare/12.1.10...12.1.11) (2022-03-24)
+
+
+### Bug Fixes
+
+* **form:** teardown subscription within the `ThyFormValidatorService` ([80d8c45](https://github.com/atinc/ngx-tethys/commit/80d8c45))
+* **menu:** remove the `click` listener once the `ThyMenuItemActionComponent` is destroyed ([bfbd661](https://github.com/atinc/ngx-tethys/commit/bfbd661))
+* **notify:** teardown the subscription once the notify container is destroyed ([d4760fc](https://github.com/atinc/ngx-tethys/commit/d4760fc))
+* **resizable:** re-enter the Angular zone when the `thyResizeStart` emits ([85e346f](https://github.com/atinc/ngx-tethys/commit/85e346f))
+* **shared:** remove the `contextmenu` listener once the `thyContextMenu` is destroyed ([7536de3](https://github.com/atinc/ngx-tethys/commit/7536de3))
+* **shared:** remove the `keydown` listener once the `ThyCtrlEnterDirective` is destroyed ([93b493f](https://github.com/atinc/ngx-tethys/commit/93b493f))
+* **shared:** remove the `keydown` listener once the `thyEnter` is destroyed ([522ddf3](https://github.com/atinc/ngx-tethys/commit/522ddf3))
+* **styles:** update default color of $success $info $warning and $danger #INFR-2752 ([#1504](https://github.com/atinc/ngx-tethys/issues/1504)) ([7bb2fc7](https://github.com/atinc/ngx-tethys/commit/7bb2fc7)), closes [#INFR-2752](https://github.com/atinc/ngx-tethys/issues/INFR-2752)
+
+
+### Features
+
+* **cascader:** #INFR-2557 activatedOptions is equal to selectedOptions during initialization ([#1506](https://github.com/atinc/ngx-tethys/issues/1506)) ([2728f49](https://github.com/atinc/ngx-tethys/commit/2728f49)), closes [#INFR-2557](https://github.com/atinc/ngx-tethys/issues/INFR-2557)
+* **layout:** add folder of layoutSidebar #INFR-2748 ([#1533](https://github.com/atinc/ngx-tethys/issues/1533)) ([7f1ffbf](https://github.com/atinc/ngx-tethys/commit/7f1ffbf)), closes [#INFR-2748](https://github.com/atinc/ngx-tethys/issues/INFR-2748)
+* **shared:** do not run change detection on the `thyStopPropagation` element and support binding changes ([ac4f8b9](https://github.com/atinc/ngx-tethys/commit/ac4f8b9))
+
+
+### Performance Improvements
+
+* **action-menu:** do not run change detection when the menu item is clicked ([6a454ec](https://github.com/atinc/ngx-tethys/commit/6a454ec))
+* **anchor:** mark `scroll` listener as passive ([15d7336](https://github.com/atinc/ngx-tethys/commit/15d7336))
+* **back-top:** do not run change detection if there are no `visibleChange` observers ([#1511](https://github.com/atinc/ngx-tethys/issues/1511)) ([5b1da11](https://github.com/atinc/ngx-tethys/commit/5b1da11))
+* **core:** do not run change detection when scrolling ([fcc1c52](https://github.com/atinc/ngx-tethys/commit/fcc1c52))
+* **dialog:** do not run change detection when the dialog is focused asynchronously ([05dbd30](https://github.com/atinc/ngx-tethys/commit/05dbd30))
+* **property-operation:** do not run change detection unnecessary when the `thy-property-operation` is clicked ([a9da5c2](https://github.com/atinc/ngx-tethys/commit/a9da5c2))
+* **resizable:** do not run change detection on mousedown、mouseenter、mouseleave and touchstart events #INFR-2501 ([b1e46c8](https://github.com/atinc/ngx-tethys/commit/b1e46c8)), closes [#INFR-2501](https://github.com/atinc/ngx-tethys/issues/INFR-2501)
+* **resizable:** mark `mousedown` and `touchstart` listeners as passive ([#1530](https://github.com/atinc/ngx-tethys/issues/1530)) ([fa91bde](https://github.com/atinc/ngx-tethys/commit/fa91bde))
+* **shared:** do not run change detection when the `thyAutofocus` focuses the element ([ee4845c](https://github.com/atinc/ngx-tethys/commit/ee4845c))
+* **shared:** run change detection only when the `thyShow` has been clicked outside ([99352f9](https://github.com/atinc/ngx-tethys/commit/99352f9))
+* **uploader:** do not run change detection when the file select is clicked ([bdb9447](https://github.com/atinc/ngx-tethys/commit/bdb9447))
+
+
+
 <a name="12.1.10"></a>
 ## [12.1.10](https://github.com/atinc/ngx-tethys/compare/12.1.9...12.1.10) (2022-03-09)